updated form-parser.php
I got rid of "requested" properties from the html file to see if "if
statements" work in the php file. When you don't fill in the required
areas like Name and Email Address, and try to submit, on the next page
you'll see  a message saying "You need to fill in all required fields."
Same thing happens when you try to submit without filling in the correct
image verification.