navigation bar practice
There are nicer ways to make a navigation bar. I commented out 3 among 4
methods in my CSS file.
1. Fade in and out
To fade the menu in and out, I used both visibility and opacity style
attributes. The reason I did that is because both carry a characteristic
that the other doesn't have. With "opacity" we can transition from
opacity: 0 to opacity: 1 and accomplish the fade effect. The problem is
that opacity: 0 doesn't remove the element. The links would still be
clickable even though they are not visible. Visibility doesn't remove
the element either, but it does remove any functionality available to
the user, so it's like it's not there.
2. Slide Down  #1
This approach sets the height of each list item in the drop down to
height: 0 and overflow: hidden, and then transitions to a set height for
list items. The overflow: hidden prevents the contents of the item to be
visible when the height is set to zero. This works beautifully until you
have a page title that is too long to fit on one line.
3. Slide Down #2
This approach uses max-height instead of height. Just like height, you
can transition to a set max-height value. Because an element with a
max-height can be any height up to and including that height, all you
have to do is set a max-height that is higher than all of your drop
downs will ever be and transition to it. This will result in
inconsistent timings for drop downs of different height to appear
because the animation is from 0 to the set max-height, not from 0 to the
height of the drop down. So if max-height: 500px, and the drop down menu
is 300 pixels tall, the timing of the transition will still be from 0 to
500.
4. Fold Out
This approach uses a combination of transitioning max-height like in the
second slide down example, but adds perspective and rotation to the
transition. perspective gives us depth, where the lower the number is
the "closer" the object appears which gives it a much more pronounced
depth effect. Higher numbers would make the effect more subtle. rotate3d
with a value of -90deg creates an effect where the element appears to
have rotated away from you. Adding a transition to all of these results
in the desired effect. Unfortunately this implementation only works with
webkit, so only Chrome and Safari users get to see it.